move: 사이드바 컴포넌트 이동 #157
