+++ frontend/src/components/Layout/Sidebar/Sidebar.style.ts
